backend/getPolarity/requirements.txt
google-cloud-language
flask
functions-framework
requests
boto3